+++ HISTORY.md
@@ -2,6 +2,7 @@ v0.1.12 (development)
 =====================
 + `indexcov`: --sex is now specified as a comma-delimited string
 + `indexcov`: handle a single sample (see #16)
++ `indexcov`: functional tests.
 
 v0.1.11
 =======